requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-registration==3.1.1
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
